third_party/blink/web_tests/external/wpt/webnn/idlharness.https.any-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': The input layout nchw is not supported."
[FAIL] NavigatorML must be primary interface of navigator
  assert_own_property: self does not have own property "NavigatorML" expected property "NavigatorML" missing
[FAIL] Stringification of navigator
  assert_class_string: class string of navigator expected "[object NavigatorML]" but got "[object Navigator]"
[FAIL] MLGraph must be primary interface of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] Stringification of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] MLGraphBuilder interface: operation constant(MLOperandDescriptor, ArrayBufferView)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation constant(double, optional MLOperandDataType)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation not(MLOperand)
  assert_own_property: interface prototype object missing non-static operation expected property "not" missing
[FAIL] MLGraphBuilder interface: builder must inherit property "not(MLOperand)" with the proper type
  assert_inherits: property "not" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling not(MLOperand) on builder with too few arguments must throw TypeError
  assert_inherits: property "not" not found in prototype chain
Harness: the test ran to completion.
